Some compatibility for DirectX. Added VC++-8.0. project files.
git-svn-id: https://spacegameengine.svn.sourceforge.net/svnroot/spacegameengine@151 469afb1b-a615-0410-ae09-9c7fe5ec7362

+++ src/plugins/d3d/src/index_buffer.cpp
@@ -94,7 +94,7 @@ void sge::d3d::index_buffer::lock(const lock_flag_t lflags, const size_type firs
 		throw std::logic_error("d3d::index_buffer::lock() you have to unlock first!");
 	void* p = 0;
 	const DWORD d3dlflags = convert_lock_flags(flags(),lflags);
-	if(buffer->Lock(first*stride, count*stride, &p, d3dlflags) != D3D_OK)
+	if(buffer->Lock(static_cast<UINT>(first * stride), static_cast<UINT>(count * stride), &p, d3dlflags) != D3D_OK)
 		throw std::runtime_error("cannot lock index buffer");
 	lock_dest = static_cast<pointer>(p);
 }
